src/requirements.txt
gym==0.24.0
klampt==0.9.0.post2
matplotlib==3.5.2
mujoco_py==2.1.2.14
numpy==1.22.3
pandas==1.4.2
PyYAML==6.0
robosuite==1.3.2
scipy==1.8.0
seaborn==0.11.2
stable_baselines3==1.5.0
transforms3d==0.3.1
